src/main/java/com/accelone/customer_management_api/service/CustomerService.java
package com.accelone.customer_management_api.service;

import com.accelone.customer_management_api.repository.CustomerRepository;
import com.accelone.customer_management_api.model.Customer;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class CustomerService {
    @Autowired
    private CustomerRepository customerRepository;

    public List<Customer> findAll() {
        return customerRepository.findAll();
    }

    public Optional<Customer> findByName(String name) {
        return customerRepository.findByName(name);
    }

    public Customer save(Customer customer) {
        return customerRepository.save(customer);
    }

    public Customer update(Customer customer) {
        return customerRepository.save(customer);
    }
}
